+++ manifest.json
@@ -1,16 +1,12 @@
 {
     "name": "Niconico Comment Grapher",
-    "version": "0.1.0",
-    "description": "Graphing the number of comments over seek bar",
+    "version": "0.2.0",
+    "description": "Graphing interval comments over seek bar",
     "permissions": [
+        "tabs",
         "webRequest",
         "webRequestBlocking",
-        "*://nicovideo.cdn.nimg.jp/web/scripts/pages/watch/watch_app_*.js",
-        "*://api.search.nicovideo.jp/api/v2/video/contents/search",
-        "tabs",
-        "*://www.nicovideo.jp/watch/*",
-        "https://api.nicoad.nicovideo.jp/",
-        "https://flapi.nicovideo.jp/"
+        "*://nicovideo.cdn.nimg.jp/web/scripts/pages/watch/watch_app_*.js"
     ],
     "content_security_policy": "script-src 'self' 'unsafe-eval'; object-src 'self'",
     "content_scripts": [{
@@ -18,17 +14,14 @@
             "*://www.nicovideo.jp/watch/*"
         ],
         "js": [
-            "js/main.js"
+            "js/index.js"
         ]
     }],
     "background": {
         "scripts": ["js/background.js"]
     },
     "web_accessible_resources": [
-        "js/hack_fetch_thread.js",
-        "js/reqAdJson.js",
-        "js/reqCmtJson.js",
-        "js/reqVideoDoc.js"
+        "js/hack_fetch_thread.js"
     ],
     "page_action": {
         "default_popup": "",
