+++ package.json
@@ -1,10 +1,12 @@
 {
   "name": "nicograph",
-  "version": "1.0.0",
+  "version": "0.2.0",
   "description": "",
   "main": "gulpfile.js",
   "dependencies": {},
   "devDependencies": {
+    "eslint": "^6.8.0",
+    "eslint-plugin-html": "^6.0.0",
     "gulp": "^4.0.2",
     "gulp-cli": "^2.2.0",
     "webpack": "^4.41.5",
@@ -13,6 +15,8 @@
     "webpack-stream": "^5.2.1"
   },
   "scripts": {
+    "clean": "npx gulp clean",
+    "build": "npx gulp prod",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
