+++ webpack.common.js
@@ -1,11 +1,8 @@
 module.exports = {
     entry: {
-        main: './src/js/main.js',
+        index: './src/js/index.js',
         hack_fetch_thread: './src/js/hack_fetch_thread.js',
-        background: './src/js/background.js',
-        reqAdJson: './src/js/reqAdJson.js',
-        reqCmtJson: './src/js/reqCmtJson.js',
-        reqVideoDoc: './src/js/reqVideoDoc.js'
+        background: './src/js/background.js'
     },
     output: {
         filename: '[name].js'
